Fix helper `isMultipleApiKey()` exported as type
