+++ model/src/data-model/index.ts
@@ -1,3 +1,4 @@
 export { InputWrapper } from '~/src/data-model/input-wrapper.js'
 export { ConditionsWrapper } from '~/src/data-model/conditions-wrapper.js'
 export { OutputType } from '~/src/data-model/enums.js'
+export { isMultipleApiKey } from '~/src/data-model/helpers.js'
